Ignore empty bins in covariance calculation
Updated bootstrapping process to assign NaNs to bins outside of the mask of that iteration. That way, the pieces inside the overall mask but outside that iteration's mask - which is precisely the bins that happened to be empty this iteration because of bootstrapping - are recorded as NaN. Then, when we average, I have it ignore NaNs.